apps/backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
openai==1.63.2
